nfs4.0/testserver.py: don't return an error when tests fail
This script was originally changed in eb3ba0b60055 ("Have testserver.py
have non-zero exit code if any tests fail"), but the same change wasn't
made to the 4.1 testserver.py script.

There also wasn't much explanation for it, and it's problematic for
running this program under the kdevops CI.

Stop the 4.0 testserver.py from exiting with an error code when any test
fails, bringing this script into line with the behavior of the 4.1
version. This makes it simpler to run under a test harness.

Cc: Frank Filz <ffilzlnx@mindspring.com>
Signed-off-by: Jeff Layton <jlayton@kernel.org>