pynfs: add CB_GETATTR tests
Just a patch to add a couple of tests for CB_GETATTR support, and
another to update the CONTRIBUTING document. I've also taken a stab at
adding support for the delstid draft to pynfs, but I'm having trouble
getting the xdr changes to build properly. I'll send a separate email
about that.

To: Calum Mackay <calum.mackay@oracle.com>
Cc: linux-nfs@vger.kernel.org
Signed-off-by: Jeff Layton <jlayton@kernel.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20240905-cb_getattr-v1-0-0af05c68234f@kernel.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20240905-cb_getattr-8db184a5b4bf",
    "prefixes": [
      "pynfs"
    ],
    "history": {
      "v1": [
        "20240905-cb_getattr-v1-0-0af05c68234f@kernel.org"
      ]
    }
  }
}